+++ src/main/java/com/trocajogo/Jogo/JogoUsuario/JogoUsuarioDTO.java
@@ -5,10 +5,11 @@ import javax.xml.bind.annotation.XmlRootElement;
 import com.trocajogo.Jogo.JogoBase;
 import com.trocajogo.Jogo.JogoPlataforma.JogoPlataformaDTO;
 
+
 @XmlRootElement(name = "JogoUsuarioDTO")
 public class JogoUsuarioDTO {
 	public int id;
-	public int idUsuario;
+	public Long idUsuario;
 	public String nomeUsuario;
 	public JogoBase jogo;
 	public JogoPlataformaDTO jogoPlataforma;
